pyproject.toml
[project]
name = "oauth-oidc"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "piyota0901", email = "piyo.tatsu.0901@gmail.com" }
]
dependencies = [
    "cryptography>=43.0.3",
    "pyjwt>=2.9.0",
    "fastapi>=0.115.3",
    "uvicorn[standard]>=0.32.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]
